refactor: [게시판] 하단 페이지 목록 컴포넌트화
